Split non-standard.idl into files per related spec (#463)
* Load non-standard.idl from a custom-idl/ directory

This is to enable splitting non-standard.idl into smaller files, and
requires a bit of refactoring of how IDL is loaded. webref.js already
had the needed logic, so it is generalized a bit.

The generated tests.json does not change at all with this change.

* Split non-standard.idl into files per related spec

The generated tests.json does not change at all with this change.