Go back to using a single service (default) and two versions (staging/production)
Updating getHost in app.js to use GAE_SERVICE instead of GAE_VERSION it
became apparent that determining the host name from the service would be
dubious. Every version in default would map to
mdn-bcd-collector.appspot.com and every version in staging would map to
staging-dot-mdn-bcd-collector.appspot.com, making it pointless to have
multiple versions.

And indeed it is. It's just a bit ugly to spell out the version in
push.yml, because it can't be specified in app.yaml. Whatever.