Make custom tests inconclusive (null) if any dependency is missing
By testing for dependencies and returning false, one could not tell from
the results if the important part of the test had run, or if it was some
precondition that had failed.

Where the code will necessarily throw an exception if the precondition
isn't met, just let the error be thrown. In the few cases where the
value isn't treated as an object, instead explicitly return null.

Related: https://github.com/foolip/mdn-bcd-collector/issues/406