Break the dependency on BCD for building tests
Instead of taking bcd.css.properties.* as input, have a list of custom
CSS properties, which is populated to match exactly the set of CSS
properties found before this change.

This is closer to what we do for IDL with custom-idl/. This makes it
more explicit what non-standard (or unmatched) CSS properties we're
generating tests for, and more importantly makes it trivial to add new
CSS properties which aren't in BCD.

A side benefit is that upgrading the BCD dependency will become much
lower risk, since it won't itself affect which tests we generate.

We will instead depend on find-missing.js to find BCD entries which we
don't have any tests for. Since non-standard CSS properties are very
rarely added these days, it should not be a big ongoing burden.

Helps with https://github.com/foolip/mdn-bcd-collector/issues/905.