Add additional custom CSS properties from Confluence (#912)
This was a mostly manual affair. Starting from
https://web-confluence.appspot.com/#!/catalog, all CSSStyleDeclaration#*
entries were converted to dashed-properties, and each property not
already covered by mdn-bcd-collector quickly checked. Members of the
CSSStyleDeclaration and CSSPropertyRule APIs showed up and were not
included.

Additionally, the layout-grid-* properties were added in both unprefixed
and -ms-prefixed forms, to be sure to cover whatever this was.

Fixes https://github.com/foolip/mdn-bcd-collector/issues/905.