Generate tests for event interfaces
This results in 812 new tests being generated for the interfaces and
their members.

The comment is true, but many event interfaces *are* exposed and it's
worth having the tests for them. Care is still needed while using the
results, but that could be added to update-bcd.js.

This reverts a change from https://github.com/foolip/mdn-bcd-collector/pull/735.