Remove unused support for update-bcd to updated prefixed entries
Prefixes and alernative names will be handled with a different approach,
with distinct tests in mdn-bcd-collector, mapping those to a subset of
the support statements in BCD. In other words, it will happen at a layer
up from most of this code, not requiring any changes to getSupportMap or
getSupportMatrix.

Because mdn-bcd-results doesn't contain any prefix information, this doesn't
change what edits are made when using the results from v2.0.1.