Once again validate custom IDL
This was dropped in https://github.com/foolip/mdn-bcd-collector/pull/959,
but that means custom IDL isn't validated at all, missing duplicate
definitions and other issues.

This doesn't add or remove any test paths, but it does change the
exposure of some tests, hopefully to be more accurate.

Tests need to be updated now that the build script requires [Exposed].