Send the right tests to SharedWorker and ServiceWorker
The tests for Worker were always sent here, which is right often enough
that nobody noticed until now. However, when there are no Worker tests,
it breaks.

This doesn't change the generated tests, but it will change which tests
are actually run, and so might have a big impact on the results.

Fixes https://github.com/foolip/mdn-bcd-collector/issues/1478.