Fix the release script (multiple bugs)
The release script currently fails using Node.js 16, because checking
out v5.0.0 and trying to install results in package-lock.json changes.
Tolerate any local changes, even though they shouldn't occur the
previous release can't be fixed when this happens.

Also do a full install, so that we're really comparing to the last
release with its actual dependencies.

Finally, there's an unmatched quote that breaks a git commit command,
and backquoted `release.js` would result in the shell trying to run that
as a command and silently failing, resulting in an empty string.