Use a single storage bucket with app version as path prefix (#1934)
The main purpose of this is to stop using GAE_VERSION, to allow it to be
anything and make it easy to deploy multiple versions in parallel.

Use the app version as a path prefix in the single bucket to clearly
separate data from different versions, and remove the need for a staging
bucket. Collisions were already unlikely, but sessions for all
deployments would still be mixed in a single bucket.

While changing things, also add /sessions/ to the path of temporary
results storage, since /files/ is already used for exported reports.