Remove custom tests for CSSConditionRule and CSSGroupingRule
These are abstract base classes that were added to the prototype chain
after some of the interfaces that inherit from them. Creating instances
for these leads to detected support before the interfaces were actually
supported.

Simple prototype testing is appropriate in this case, although it's
possible some browser supported the interfaces without exposing the
interface objects. Regardless of how this is testing, it will need to be
manually scrutinized when updating the data.