Add intended test and reorganize
Prior to this commit, the expected error in a test for the
`inferSupportStatements` function was satisfied by a similar error from
the `getSupportMatrix` function.

Modify the test to demonstrate the intended error. Add a second test to
preserve coverage for `getSupportMatrix`. In both cases, define the
expected error more precisely in order to reduce risk for similar false
positives in the future.