bcd-update: do not spuriously report modifications (#2326)
Because the return value of `update` is used only to optimize I/O, false
positives only degrade efficiency--not correctness. However, the
underlying behavior substantively effects the interpretation of the
`"mirror"` support value. Due to an unrelated bug (where modifications
to mirrored support values are not written back to disk), this error
cannot currently be observed.

Correct the return value for this condition in order to improve
efficiency and prepare for a later correction to the `"mirror"` feature.