Formalize version range handling (#2353)
Prior to this commit, the `version_added` property could take values
which were undesirable for display to end-users (namely, those including
the substring `0>`). Care was taken to avoid writing strings of this
form to browser compatibility data structures.

Refactor the internals to preclude undesirable values, simplifying the
code necessary to update the data structure and eliminating the risk of
regression in future code paths.